backend/api_requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
cloud-sql-python-connector[asyncpg]==1.12.1
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.2
google-cloud-bigquery==3.25.0
pyfarmhash==0.3.2
httpx==0.27.2
